Add VCL Protocol Decoder
Note this protocol has no End Byte and no Checksum and is variable length. This means the VCL detector will often detect VCL protocol when fed other telemetry formats. TODO - Watch for 3 VCL commands back to back (with no unexpected bytes ibetween) to confirm it is VCL data and then stop decoding if it is clearly not VCL